feat: set carrousel to hook
