+++ srcs/components/Projects/Carrousel/Carrousel.style.tsx
@@ -3,11 +3,12 @@ import styled from 'styled-components'
 export const Layout = styled.div<{ width: number}>`
     width: 90%;
     height: 80vh;
-
+    
     margin-top: 3rem;
     
     overflow: hidden;
 
+    border: 1px solid black;
     position: relative;
 `
 export const Slide = styled.div<{ width: number, num: number, idx: number, duration: number }>`
