+++ srcs/hooks/index.tsx
@@ -1,2 +1,3 @@
 export { default as useCanvas } from './useCanvas'
 export { default as useTypoGraph } from './useTypoGraph'
+export { default as useCarrousel } from './useCarrousel'
